safari-600.7-branch/LayoutTests/svg/dom/SVGAnimatedBoolean-expected.txt

This test checks the SVGAnimatedBoolean API - utilizing the externalResourcesRequired property of SVGRectElement

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Check initial SVGExternalResourcesRequired value
PASS rectElement.externalResourcesRequired.baseVal is false

Set value to true
PASS rectElement.externalResourcesRequired.baseVal = true is true

Caching baseVal in local variable
PASS baseVal is true

Modify local baseVal variable to true
PASS baseVal = false is false

Assure that rectElement.externalResourcesRequired has not been changed, but the local baseVal variable
PASS baseVal is false
PASS rectElement.externalResourcesRequired.baseVal is true

Check assigning values of various types
PASS rectElement.externalResourcesRequired.baseVal = rectElement.externalResourcesRequired is rectElement.externalResourcesRequired
PASS rectElement.externalResourcesRequired.baseVal is true
PASS rectElement.externalResourcesRequired.baseVal = null is null
PASS rectElement.externalResourcesRequired.baseVal is false
PASS rectElement.externalResourcesRequired.baseVal = 'aString' is 'aString'
PASS rectElement.externalResourcesRequired.baseVal is true
PASS rectElement.externalResourcesRequired.baseVal = rectElement is rectElement
PASS rectElement.externalResourcesRequired.baseVal is true
PASS successfullyParsed is true

TEST COMPLETE
